Change uint64 to uint32 for count/metric fields
Protobuf JSON marshaling encodes uint64 as strings but uint32 as numbers.
This change converts count and metric fields to uint32 so they marshal as
JSON numbers by default, which is more natural for clients consuming the API.

Fields changed to uint32:
- `Scale`: MinInstances, MaxInstances, TargetCPUUsageMilli, TargetMemoryUsageMiB, TargetInFlightRequests
- `ScaleDecision`: DesiredInstances, UnclampedDesiredInstances
- `Instance`: CPUUsageMilli, MemoryUsageMiB
- `Heartbeat`: InFlightRequests

`FunctionHash` remains uint64 as it's internal and needs the full hash space.

The `MarshalJSON` methods no longer need `json.StringifyNumbers(true)` since
uint32 marshals as numbers by default. The `UnmarshalJSON` methods still
accept both string and number formats for backwards compatibility during
rolling deployments.

+++ internal/controller/client.go
@@ -17,7 +17,7 @@ import (
 type Client interface {
 	Instance(ctx context.Context, fn *skipper.Function, excludeInstanceNames ...string) (instance *skipper.Instance, err error)
 	Heartbeat(ctx context.Context, routerIP string, heartbeats []*skipper.Heartbeat, forwardedFor ...string) error
-	Scale(ctx context.Context, fn *skipper.Function, desiredInstances uint64, reason skipper.ScaleReason) ([]*skipper.Instance, error)
+	Scale(ctx context.Context, fn *skipper.Function, desiredInstances uint32, reason skipper.ScaleReason) ([]*skipper.Instance, error)
 }
 
 type NewClientFunc func(host string, port int) Client
@@ -99,14 +99,14 @@ func (c *httpClient) Heartbeat(ctx context.Context, routerIP string, heartbeats 
 	return nil
 }
 
-func (c *httpClient) Scale(ctx context.Context, fn *skipper.Function, desiredInstances uint64, reason skipper.ScaleReason) ([]*skipper.Instance, error) {
+func (c *httpClient) Scale(ctx context.Context, fn *skipper.Function, desiredInstances uint32, reason skipper.ScaleReason) ([]*skipper.Instance, error) {
 	req, err := http.NewRequestWithContext(ctx, http.MethodPost, c.addr+"/scale", nil)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create scale request: %w", err)
 	}
 
 	fn.SetHeader(req)
-	req.Header[key.DesiredInstances.Header] = []string{strconv.FormatUint(desiredInstances, 10)}
+	req.Header[key.DesiredInstances.Header] = []string{strconv.FormatUint(uint64(desiredInstances), 10)}
 	req.Header[key.Reason.Header] = []string{string(reason)}
 
 	res, err := c.Do(req)
